+++ CHANGES.rst
@@ -4,6 +4,9 @@
 - updatewcs() now reads all extension immediately after opening a file
   to fix a problem after astropy implemented fits lazy loading. [#21]
 
+- Fixed a bug in updating the D2IM correction in a science file when the 
+  a new distortion file was supplied through D2IMFILE keyword. [#22]
+
 1.2.4 (2016-10-27)
 ------------------